JP2KAK: avoid int32 overflow on creation of images with more than 2 billion pixels
git-svn-id: https://svn.osgeo.org/gdal/trunk@39923 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/jp2kak/jp2kakdataset.cpp
@@ -2146,8 +2146,8 @@ JP2KAKCreateCopy( const char * pszFilename, GDALDataset *poSrcDS,
 
     // Avoid splitting into too many tiles - apparently limiting to 64K tiles.
     // There is a hard limit on the number of tiles allowed in JPEG2000.
-    const double dfXbyY = static_cast<double>(nXSize * nYSize) / (1024 * 64);
-    while( dfXbyY >= static_cast<double>(nTileXSize * nTileYSize) )
+    const double dfXbyY = static_cast<double>(nXSize) * nYSize / (1024 * 64);
+    while( dfXbyY >= static_cast<double>(nTileXSize) * nTileYSize )
     {
         nTileXSize *= 2;
         nTileYSize *= 2;
@@ -2575,7 +2575,7 @@ JP2KAKCreateCopy( const char * pszFilename, GDALDataset *poSrcDS,
             // Setup scaled progress monitor.
 
             const double dfPixelsDoneAfter =
-                dfPixelsDone + (nThisTileXSize * nThisTileYSize);
+                dfPixelsDone + static_cast<double>(nThisTileXSize) * nThisTileYSize;
 
             void *pScaledProgressData = GDALCreateScaledProgress(
                 dfPixelsDone / dfPixelsTotal, dfPixelsDoneAfter / dfPixelsTotal,
